feat: Implement Best first search

+++ src/heuristic/Heuristic.java
@@ -3,5 +3,5 @@ package heuristic;
 import state.Cubie;
 
 public interface Heuristic {
-    long heuristic(Cubie state);
+    int heuristic(Cubie state);
 }
